+++ CHANGELOG.md
@@ -64,6 +64,13 @@ declined or not answered. Currently, the server accepts all incoming calls, play
     - Parameter `module::PushNotification/windowsphone-package-sid` (deprecated in 2.3.0)
     - Parameter `module::PushNotification/windowsphone-application-secret` (deprecated in 2.3.0)
     - Silk audio encoder/decoder is not supported anymore.
+    - Parameter `module::Router/stateful` (deprecated in 2.0.0)
+    - Parameter `module::Router/fork` (deprecated in 2.0.0)
+    - Parameter `module::Router/generated-contact-route` (deprecated in 2.0.0)
+    - Parameter `module::Router/generated-contact-expected-realm` (deprecated in 2.0.0)
+    - Parameter `module::Router/generate-contact-even-on-filled-aor` (deprecated in 2.0.0)
+    - Parameter `module::Router/preroute` (deprecated in 2.0.0)
+    - Parameter `module::Router/call-push-response-timeout` (deprecated in 2.2.0)
 - **Conference:**
     - Parameter `conference-server/conference-factory-uri` (deprecated in 2.1.0)
 - **EventLogs:**